+++ sentry-android-core/src/main/java/io/sentry/android/core/ActivityLifecycleIntegration.java
@@ -9,8 +9,6 @@ import android.content.Context;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Process;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import io.sentry.Breadcrumb;
 import io.sentry.IHub;
 import io.sentry.ISpan;
@@ -28,6 +26,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.WeakHashMap;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.TestOnly;
 import org.jetbrains.annotations.VisibleForTesting;
 
@@ -116,7 +115,7 @@ public final class ActivityLifecycleIntegration
     activityFramesTracker.stop();
   }
 
-  private void addBreadcrumb(final @NonNull Activity activity, final @NotNull String state) {
+  private void addBreadcrumb(final @NotNull Activity activity, final @NotNull String state) {
     if (options != null && hub != null && options.isEnableActivityLifecycleBreadcrumbs()) {
       final Breadcrumb breadcrumb = new Breadcrumb();
       breadcrumb.setType("navigation");
@@ -128,7 +127,7 @@ public final class ActivityLifecycleIntegration
     }
   }
 
-  private @NotNull String getActivityName(final @NonNull Activity activity) {
+  private @NotNull String getActivityName(final @NotNull Activity activity) {
     return activity.getClass().getSimpleName();
   }
 
@@ -140,7 +139,7 @@ public final class ActivityLifecycleIntegration
     }
   }
 
-  private void startTracing(final @NonNull Activity activity) {
+  private void startTracing(final @NotNull Activity activity) {
     if (performanceEnabled && !isRunningTransaction(activity) && hub != null) {
       // as we allow a single transaction running on the bound Scope, we finish the previous ones
       stopPreviousTransactions();
@@ -151,9 +150,10 @@ public final class ActivityLifecycleIntegration
 
       final Date appStartTime =
           foregroundImportance ? AppStartState.getInstance().getAppStartTime() : null;
+      final Boolean coldStart = AppStartState.getInstance().isColdStart();
 
       // in case appStartTime isn't available, we don't create a span for it.
-      if (firstActivityCreated || appStartTime == null) {
+      if (firstActivityCreated || appStartTime == null || coldStart == null) {
         transaction =
             hub.startTransaction(
                 activityName,
@@ -176,7 +176,9 @@ public final class ActivityLifecycleIntegration
                 });
         // start specific span for app start
 
-        appStartSpan = transaction.startChild(getAppStartOp(), getAppStartDesc(), appStartTime);
+        appStartSpan =
+            transaction.startChild(
+                getAppStartOp(coldStart), getAppStartDesc(coldStart), appStartTime);
       }
 
       // lets bind to the scope so other integrations can pick it up
@@ -208,11 +210,11 @@ public final class ActivityLifecycleIntegration
         });
   }
 
-  private boolean isRunningTransaction(final @NonNull Activity activity) {
+  private boolean isRunningTransaction(final @NotNull Activity activity) {
     return activitiesWithOngoingTransactions.containsKey(activity);
   }
 
-  private void stopTracing(final @NonNull Activity activity, final boolean shouldFinishTracing) {
+  private void stopTracing(final @NotNull Activity activity, final boolean shouldFinishTracing) {
     if (performanceEnabled && shouldFinishTracing) {
       final ITransaction transaction = activitiesWithOngoingTransactions.get(activity);
       finishTransaction(transaction);
@@ -238,7 +240,7 @@ public final class ActivityLifecycleIntegration
 
   @Override
   public synchronized void onActivityCreated(
-      final @NonNull Activity activity, final @Nullable Bundle savedInstanceState) {
+      final @NotNull Activity activity, final @Nullable Bundle savedInstanceState) {
     setColdStart(savedInstanceState);
     addBreadcrumb(activity, "created");
     startTracing(activity);
@@ -247,7 +249,7 @@ public final class ActivityLifecycleIntegration
   }
 
   @Override
-  public synchronized void onActivityStarted(final @NonNull Activity activity) {
+  public synchronized void onActivityStarted(final @NotNull Activity activity) {
     // The docs on the screen rendering performance tracing
     // (https://firebase.google.com/docs/perf-mon/screen-traces?platform=android#definition),
     // state that the tracing starts for every Activity class when the app calls .onActivityStarted.
@@ -259,8 +261,8 @@ public final class ActivityLifecycleIntegration
   }
 
   @Override
-  public synchronized void onActivityResumed(final @NonNull Activity activity) {
-    if (!firstActivityResumed && performanceEnabled) {
+  public synchronized void onActivityResumed(final @NotNull Activity activity) {
+    if (!firstActivityResumed) {
 
       // we only finish the app start if the process is of foregroundImportance
       if (foregroundImportance) {
@@ -277,7 +279,7 @@ public final class ActivityLifecycleIntegration
       }
 
       // finishes app start span
-      if (appStartSpan != null) {
+      if (performanceEnabled && appStartSpan != null) {
         appStartSpan.finish();
       }
       firstActivityResumed = true;
@@ -292,7 +294,7 @@ public final class ActivityLifecycleIntegration
   }
 
   @Override
-  public synchronized void onActivityPostResumed(final @NonNull Activity activity) {
+  public synchronized void onActivityPostResumed(final @NotNull Activity activity) {
     // only executed if API >= 29 otherwise it happens on onActivityResumed
     if (isAllActivityCallbacksAvailable && options != null) {
       // this should be called only when onResume has been executed already, which means
@@ -302,23 +304,23 @@ public final class ActivityLifecycleIntegration
   }
 
   @Override
-  public synchronized void onActivityPaused(final @NonNull Activity activity) {
+  public synchronized void onActivityPaused(final @NotNull Activity activity) {
     addBreadcrumb(activity, "paused");
   }
 
   @Override
-  public synchronized void onActivityStopped(final @NonNull Activity activity) {
+  public synchronized void onActivityStopped(final @NotNull Activity activity) {
     addBreadcrumb(activity, "stopped");
   }
 
   @Override
   public synchronized void onActivitySaveInstanceState(
-      final @NonNull Activity activity, final @NonNull Bundle outState) {
+      final @NotNull Activity activity, final @NotNull Bundle outState) {
     addBreadcrumb(activity, "saveInstanceState");
   }
 
   @Override
-  public synchronized void onActivityDestroyed(final @NonNull Activity activity) {
+  public synchronized void onActivityDestroyed(final @NotNull Activity activity) {
     addBreadcrumb(activity, "destroyed");
 
     // in case the appStartSpan isn't completed yet, we finish it as cancelled to avoid
@@ -355,23 +357,23 @@ public final class ActivityLifecycleIntegration
   }
 
   private void setColdStart(final @Nullable Bundle savedInstanceState) {
-    if (!firstActivityCreated && performanceEnabled) {
+    if (!firstActivityCreated) {
       // if Activity has savedInstanceState then its a warm start
       // https://developer.android.com/topic/performance/vitals/launch-time#warm
       AppStartState.getInstance().setColdStart(savedInstanceState == null);
     }
   }
 
-  private @NotNull String getAppStartDesc() {
-    if (AppStartState.getInstance().isColdStart()) {
+  private @NotNull String getAppStartDesc(final boolean coldStart) {
+    if (coldStart) {
       return "Cold Start";
     } else {
       return "Warm Start";
     }
   }
 
-  private @NotNull String getAppStartOp() {
-    if (AppStartState.getInstance().isColdStart()) {
+  private @NotNull String getAppStartOp(final boolean coldStart) {
+    if (coldStart) {
       return APP_START_COLD;
     } else {
       return APP_START_WARM;