+++ sentry-android-fragment/src/main/java/io/sentry/android/fragment/SentryFragmentLifecycleCallbacks.kt
@@ -161,6 +161,6 @@ class SentryFragmentLifecycleCallbacks(
     }
 
     companion object {
-        const val FRAGMENT_LOAD_OP = "ui.fragment.load"
+        const val FRAGMENT_LOAD_OP = "ui.load"
     }
 }