+++ sentry/src/test/java/io/sentry/protocol/DeviceTest.kt
@@ -61,6 +61,7 @@ class DeviceTest {
         device.language = "language"
         device.connectionType = "connection type"
         device.batteryTemperature = 30f
+        device.locale = "en-US"
         val unknown = mapOf(Pair("unknown", "unknown"))
         device.setUnknown(unknown)
 
@@ -98,6 +99,7 @@ class DeviceTest {
         assertEquals("language", clone.language)
         assertEquals("connection type", clone.connectionType)
         assertEquals(30f, clone.batteryTemperature)
+        assertEquals("en-US", clone.locale)
         assertNotNull(clone.unknown) {
             assertEquals("unknown", it["unknown"])
         }